requirements.txt
openai==1.93.0
anthropic==0.55.0
elasticsearch==8.16.0
python-dotenv==1.0.0
pydantic==2.11.7
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pymemcache==4.0.0
numpy==2.2.4
aiohttp==3.12.13
boto3==1.37.34
requests==2.32.3
sentry-sdk==2.41.0